feat: extract phw specific rules to own file; extract validation exception for reuse
