+++ hl7_server/hl7_server/exceptions/validation_exception.py
@@ -1,0 +1,2 @@
+class ValidationException(Exception):
+    pass
